backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.26.0

# LangChain
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.18

# Token counting
tiktoken==0.5.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3

# Code quality
black==24.1.1
ruff==0.1.15
